+++ src/scene_server/admin_server/imports.go
@@ -128,5 +128,5 @@ import (
 	_ "configcenter/src/scene_server/admin_server/upgrader/y3.9.202106291420"
 	_ "configcenter/src/scene_server/admin_server/upgrader/y3.9.202106301910"
 	_ "configcenter/src/scene_server/admin_server/upgrader/y3.9.202107011154"
-	_ "configcenter/src/scene_server/admin_server/upgrader/y3.9.202107012030"
+	_ "configcenter/src/scene_server/admin_server/upgrader/y3.9.202107142030"
 )
